change uint64_t to int64_t for all

+++ stablehlo/dialect/TypeInference.cpp
@@ -595,13 +595,13 @@ LogicalResult verifyReducerShape(Optional<Location> loc, Block& block,
 
 LogicalResult verifyReduceScatter(Optional<Location> location,
                                   TypeRange operandTypes, TypeRange resultTypes,
-                                  uint64_t scatterDimension) {
+                                  int64_t scatterDimension) {
   // If operand and result are both ranked, then the size of the scatter
   // dimension in the operand should be a multiple of the size of the scatter
   // dimension in the result.
 
   // TODO(zhouxin) Change the ODS definition to return int64_t.
-  if (static_cast<int64_t>(scatterDimension) < 0) {
+  if (scatterDimension < 0) {
     return emitOptionalError(location, "expects scatter_dimension >= 0");
   }
 
@@ -612,7 +612,7 @@ LogicalResult verifyReduceScatter(Optional<Location> location,
     if (operandType.getRank() != resultType.getRank())
       return emitOptionalError(location,
                                "operand and result should have same rank");
-    if (static_cast<int64_t>(scatterDimension) >= operandType.getRank())
+    if (scatterDimension >= operandType.getRank())
       return emitOptionalError(
           location, "scatter dim should be less than operand/result rank");
     if (operandType.isDynamicDim(scatterDimension) ||
@@ -634,8 +634,8 @@ LogicalResult verifyReduceScatter(Optional<Location> location,
 
     // Non scatter dimensions should be equal.
     for (uint64_t index : llvm::seq<uint64_t>(0, operandType.getRank())) {
-      if (index == scatterDimension || operandType.isDynamicDim(index) ||
-          resultType.isDynamicDim(index))
+      if (static_cast<int64_t>(index) == scatterDimension ||
+          operandType.isDynamicDim(index) || resultType.isDynamicDim(index))
         continue;
       if (operandType.getDimSize(index) != resultType.getDimSize(index))
         return emitOptionalError(
@@ -824,7 +824,7 @@ LogicalResult inferAllToAllOp(
 
 LogicalResult inferBatchNormGradOp(
     Optional<Location> location, Value operand, Value scale,
-    uint64_t featureIndex,
+    int64_t featureIndex,
     SmallVectorImpl<ShapedTypeComponents>& inferredReturnShapes) {
   if (failed(verifyBatchNorm(location, operand, scale, featureIndex)))
     return failure();
@@ -840,7 +840,7 @@ LogicalResult inferBatchNormGradOp(
 
 LogicalResult inferBatchNormInferenceOp(
     Optional<Location> location, Value operand, Value scale,
-    uint64_t featureIndex,
+    int64_t featureIndex,
     SmallVectorImpl<ShapedTypeComponents>& inferredReturnShapes) {
   if (failed(verifyBatchNorm(location, operand, scale, featureIndex)))
     return failure();
@@ -851,7 +851,7 @@ LogicalResult inferBatchNormInferenceOp(
 
 LogicalResult inferBatchNormTrainingOp(
     Optional<Location> location, Value operand, Value scale,
-    uint64_t featureIndex,
+    int64_t featureIndex,
     SmallVectorImpl<ShapedTypeComponents>& inferredReturnShapes) {
   if (failed(verifyBatchNorm(location, operand, scale, featureIndex)))
     return failure();
@@ -2220,15 +2220,14 @@ LogicalResult verifyDynamicReshapeOp(Optional<Location> location,
   return success();
 }
 
-LogicalResult verifyIotaOp(Optional<Location> location, uint64_t iotaDimension,
+LogicalResult verifyIotaOp(Optional<Location> location, int64_t iotaDimension,
                            Type resultType) {
   auto shape = resultType.cast<ShapedType>();
   if (!shape.hasRank()) return success();
   if (shape.getRank() == 0)
     return emitOptionalError(location, "does not support scalars.");
 
-  if (static_cast<int64_t>(iotaDimension) >= shape.getRank() ||
-      static_cast<int64_t>(iotaDimension) < 0)
+  if (iotaDimension >= shape.getRank() || iotaDimension < 0)
     return emitOptionalError(
         location,
         "iota dimension cannot go beyond the output rank or be negative.");
@@ -2308,7 +2307,7 @@ LogicalResult verifyReduceOp(Optional<Location> location, ValueRange inputs,
 }
 
 LogicalResult verifyReduceScatterOp(Optional<Location> location, Value operand,
-                                    uint64_t scatterDimension,
+                                    int64_t scatterDimension,
                                     DenseIntElementsAttr replicaGroups,
                                     bool useGlobalDeviceIds,
                                     Region& computation,
@@ -2381,7 +2380,7 @@ LogicalResult verifyReduceWindowOp(
 }
 
 LogicalResult verifySortOp(Optional<Location> location, ValueRange inputs,
-                           uint64_t dimension, Region& comparator) {
+                           int64_t dimension, Region& comparator) {
   auto operandTypes = inputs.getTypes();
   for (auto operandType : operandTypes) {
     auto operandShapedType = operandType.cast<ShapedType>();
